requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
defusedxml==0.7.1
dnspython==2.7.0
frozenlist==1.6.0
geoip2==5.0.1
idna==3.10
ipwhois==1.3.0
maxminddb==2.6.3
multidict==6.4.3
propcache==0.3.1
requests==2.32.3
urllib3==2.4.0
yarl==1.20.0
